# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorMessage": {
        "properties": {
          "message": {
            "description": "What went wrong, in words to show the user.",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "type": "object"
      },
...
      "Message": {
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "supabase": {
        "bearerFormat": "JWT",
        "description": "A Supabase access token for the signed-in user.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "The Electriplan web app's API. Every endpoint needs a Supabase access token (`Authorization: Bearer ...`). Endpoints that work inside one company take the `X-Organisation-Id` header and say which permission they need; errors come back as `{\"message\": \"...\"}`.\n\nThe floor-plan analyser (`/api/floorplan/...`) is a separate service and is not described here.",
    "title": "Electriplan API",
    "version": "v1"
  },
  "openapi": "3.0.1",
  "paths": {
    "/api/files/{id}": {
      "get": {
        "description": "The bytes, with their type. A file's contents never change, so it may be cached.\n\n**Permission:** `company.view` (roles: owner, admin, builder, electrician, viewer).",
        "operationId": "getFile",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "The company to act in. Needed only when the caller belongs to more than one.",
            "in": "header",
            "name": "X-Organisation-Id",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/*": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "The file"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "description": "No company chosen while the caller belongs to several, or the header is not a company id."
          },
          "401": {
            "description": "Not signed in, or the access token is invalid or expired."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            },
            "description": "The caller is not an active member of that company, or it is closed. Or the caller's role there lacks `company.view`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "No such file in this company."
          }
        },
        "summary": "A file of the company",
        "tags": [
          "Files"
        ],
        "x-permission": "company.view",
        "x-roles": [
          "owner",
          "admin",
          "builder",
          "electrician",
          "viewer"
        ]
      }
    },
...
  },
  "security": [
    {
      "supabase": []
    }
  ],
  "servers": [
    {
      "description": "The web app's origin (it proxies /api to this service)",
      "url": "/"
    }
  ],
...
}